LayoutTests/imported/w3c/web-platform-tests/trusted-types/block-string-assignment-to-attribute-via-attribute-node-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.



PASS Sanity check: Setting non-TT attributes still works.
FAIL Set embed.src via textContent assert_throws_js: function "_ => {
      element.attributes[0].textContent = "sldkjsfldk";
    }" did not throw
FAIL Set embed.src via nodeValue assert_throws_js: function "_ => {
      element.attributes[0].nodeValue = "sdflkgjdlkgjdg";
    }" did not throw
FAIL Set object.data via textContent assert_throws_js: function "_ => {
      element.attributes[0].textContent = "sldkjsfldk";
    }" did not throw
FAIL Set object.data via nodeValue assert_throws_js: function "_ => {
      element.attributes[0].nodeValue = "sdflkgjdlkgjdg";
    }" did not throw
FAIL Set object.codebase via textContent assert_throws_js: function "_ => {
      element.attributes[0].textContent = "sldkjsfldk";
    }" did not throw
FAIL Set object.codebase via nodeValue assert_throws_js: function "_ => {
      element.attributes[0].nodeValue = "sdflkgjdlkgjdg";
    }" did not throw
FAIL Set script.src via textContent assert_throws_js: function "_ => {
      element.attributes[0].textContent = "sldkjsfldk";
    }" did not throw
FAIL Set script.src via nodeValue assert_throws_js: function "_ => {
      element.attributes[0].nodeValue = "sdflkgjdlkgjdg";
    }" did not throw
FAIL Set iframe.srcdoc via textContent assert_throws_js: function "_ => {
      element.attributes[0].textContent = "sldkjsfldk";
    }" did not throw
FAIL Set iframe.srcdoc via nodeValue assert_throws_js: function "_ => {
      element.attributes[0].nodeValue = "sdflkgjdlkgjdg";
    }" did not throw
FAIL Set div.onclick via textContent assert_throws_js: function "_ => {
      element.attributes[0].textContent = "sldkjsfldk";
    }" did not throw
FAIL Set div.onclick via nodeValue assert_throws_js: function "_ => {
      element.attributes[0].nodeValue = "sdflkgjdlkgjdg";
    }" did not throw
